+++ test/pow_assent/ecto/schema_test.exs
@@ -60,29 +60,82 @@ defmodule PowAssent.Ecto.SchemaTest do
       assert [user_identity] = changeset.changes.user_identities
       assert user_identity.errors[:uid_provider] == {"has already been taken", [constraint: :unique, constraint_name: "user_identities_uid_provider_index"]}
     end
+  end
 
+  defmodule UsernameUserWithEmail do
+    @moduledoc false
+    use Ecto.Schema
+    use Pow.Ecto.Schema, user_id_field: :username
+    use Pow.Extension.Ecto.Schema,
+      extensions: [PowEmailConfirmation]
+    use PowAssent.Ecto.Schema
+
+    schema "users" do
+      has_many :user_identities, PowAssent.Test.Ecto.UserIdentities.UserIdentity, foreign_key: :user_id, on_delete: :delete_all
+
+      field :email, :string
+
+      pow_user_fields()
+
+      timestamps()
+    end
+
+    def user_identity_changeset(user_or_changeset, user_identity, attrs, user_id_attrs) do
+      user_or_changeset
+      |> Ecto.Changeset.cast(attrs, [:email])
+      |> pow_assent_user_identity_changeset(user_identity, attrs, user_id_attrs)
+    end
+  end
+
+  describe "user_identity_changeset/4 with PowEmailConfirmation" do
     test "sets :email_confirmed_at when provided as attrs" do
       provider_params = %{email: "test@example.com", email_verified: true, name: "John Doe"}
 
-      changeset = UserConfirmEmail.user_identity_changeset(%UserConfirmEmail{}, @user_identity, provider_params, %{email: "test@example.com"})
+      changeset = User.user_identity_changeset(%User{}, @user_identity, provider_params, nil)
       assert changeset.changes[:email]
       refute changeset.changes[:email_confirmed_at]
+      refute changeset.changes[:email_confirmation_token]
+
+      changeset = UserConfirmEmail.user_identity_changeset(%UserConfirmEmail{}, @user_identity, provider_params, %{email: "foo@example.com"})
+      assert changeset.changes[:email]
+      refute changeset.changes[:email_confirmed_at]
+      assert changeset.changes[:email_confirmation_token]
+
+      changeset = UserConfirmEmail.user_identity_changeset(%UserConfirmEmail{}, @user_identity, provider_params, %{email: "test@example.com"})
+      assert changeset.changes[:email]
+      assert changeset.changes[:email_confirmed_at]
+      refute changeset.changes[:email_confirmation_token]
 
       changeset = UserConfirmEmail.user_identity_changeset(%UserConfirmEmail{}, @user_identity, provider_params, nil)
       assert changeset.changes[:email]
       assert changeset.changes[:name] == "John Doe"
       assert changeset.changes[:email_confirmed_at]
+      refute changeset.changes[:email_confirmation_token]
 
       changeset = UserConfirmEmail.user_identity_changeset(%UserConfirmEmail{}, @user_identity,  Map.delete(provider_params, :email_verified), nil)
       assert changeset.changes[:email]
       assert changeset.changes[:name] == "John Doe"
       refute changeset.changes[:email_confirmed_at]
-
-      changeset = User.user_identity_changeset(%User{}, @user_identity, provider_params, nil)
-      assert changeset.changes[:email]
-      refute changeset.changes[:email_confirmed_at]
+      assert changeset.changes[:email_confirmation_token]
     end
 
+    test "sets :email_confirmed_at when provided as attrs and :email is not user id field" do
+      provider_params = %{email: "test@example.com", email_verified: true}
+
+      changeset = UsernameUserWithEmail.user_identity_changeset(%UsernameUserWithEmail{}, @user_identity, Map.delete(provider_params, :email_verified), %{username: "john.doe"})
+      assert changeset.changes[:email]
+      refute changeset.changes[:email_confirmed_at]
+      assert changeset.changes[:email_confirmation_token]
+
+      changeset = UsernameUserWithEmail.user_identity_changeset(%UsernameUserWithEmail{}, @user_identity, provider_params, %{username: "john.doe"})
+      assert changeset.changes[:username]
+      assert changeset.changes[:email]
+      assert changeset.changes[:email_confirmed_at]
+      refute changeset.changes[:email_confirmation_token]
+    end
+  end
+
+  describe "user_identity_changeset/4 with PowInvitation" do
     test "sets :invitation_accepted_at when is invited user" do
       changeset = InvitationUser.user_identity_changeset(%InvitationUser{}, @user_identity, %{}, %{email: "test@example.com"})
       refute changeset.changes[:invitation_accepted_at]