+++ test/pow_assent/phoenix/controllers/registration_controller_test.exs
@@ -5,10 +5,16 @@ defmodule PowAssent.Phoenix.RegistrationControllerTest do
   @token_params %{"access_token" => "access_token"}
   @user_identity_params %{"provider" => @provider, "uid" => "new_user", "token" => @token_params}
   @user_params %{"name" => "John Doe"}
-  @provider_params %{@provider => %{user_identity: @user_identity_params, user: @user_params}}
+
+  defp provider_params(opts \\ []) do
+    user_identity_params = Map.merge(@user_identity_params, Keyword.get(opts, :user_identity_params, %{}))
+    user_params = Map.merge(@user_params, Keyword.get(opts, :user_params, %{}))
+
+    %{@provider => %{user_identity: user_identity_params, user: user_params}}
+  end
 
   setup %{conn: conn} do
-    conn = Plug.Conn.put_session(conn, :pow_assent_params, @provider_params)
+    conn = Plug.Conn.put_session(conn, :pow_assent_params, provider_params())
 
     {:ok, conn: conn}
   end
@@ -25,7 +31,7 @@ defmodule PowAssent.Phoenix.RegistrationControllerTest do
     end
 
     test "with invalid provider path", %{conn: conn} do
-      conn = get(conn, Routes.pow_assent_registration_path(conn, :add_user_id, "invalid"))
+      conn = get conn, Routes.pow_assent_registration_path(conn, :add_user_id, "invalid")
 
       assert redirected_to(conn) == "/logged-out"
       assert get_flash(conn, :error) == "Invalid Request."
@@ -41,18 +47,21 @@ defmodule PowAssent.Phoenix.RegistrationControllerTest do
   end
 
   describe "POST /auth/:provider/create" do
+    @valid_params %{user: %{email: "foo@example.com"}}
+    @taken_params %{user: %{email: "taken@example.com"}}
+
     test "with missing session params", %{conn: conn} do
       conn =
         conn
         |> Plug.Conn.delete_session(:pow_assent_params)
-        |> post(Routes.pow_assent_registration_path(conn, :create, @provider), %{user: %{email: "foo@example.com"}})
+        |> post(Routes.pow_assent_registration_path(conn, :create, @provider), @valid_params)
 
       assert redirected_to(conn) == "/logged-out"
       assert get_flash(conn, :error) == "Invalid Request."
     end
 
     test "with valid params", %{conn: conn} do
-      conn = post conn, Routes.pow_assent_registration_path(conn, :create, @provider), %{user: %{email: "foo@example.com"}}
+      conn = post conn, Routes.pow_assent_registration_path(conn, :create, @provider), @valid_params
 
       assert redirected_to(conn) == "/registration_created"
       assert user = Pow.Plug.current_user(conn)
@@ -60,8 +69,8 @@ defmodule PowAssent.Phoenix.RegistrationControllerTest do
       assert get_flash(conn, :info) == "user_created_test_provider"
     end
 
-    test "with invalid params", %{conn: conn} do
-      conn = post conn, Routes.pow_assent_registration_path(conn, :create, @provider), %{user: %{email: "taken@example.com"}}
+    test "with taken user id params", %{conn: conn} do
+      conn = post conn, Routes.pow_assent_registration_path(conn, :create, @provider), @taken_params
 
       assert html = html_response(conn, 200)
       assert html =~ "<label for=\"user_email\">Email</label>"
@@ -70,11 +79,11 @@ defmodule PowAssent.Phoenix.RegistrationControllerTest do
     end
 
     test "with identity already bound to another user", %{conn: conn} do
-      params = %{@provider => %{user_identity: %{"provider" => @provider, "uid" => "identity_taken", "token" => @token_params}, user: %{"name" => "John Doe"}}}
-      conn =
+      params = provider_params(user_identity_params: %{"uid" => "identity_taken"})
+      conn   =
         conn
         |> Plug.Conn.put_session(:pow_assent_params, params)
-        |> post(Routes.pow_assent_registration_path(conn, :create, @provider), %{user: %{email: "foo@example.com"}})
+        |> post(Routes.pow_assent_registration_path(conn, :create, @provider), @valid_params)
 
       assert redirected_to(conn) == Routes.pow_registration_path(conn, :new)
       assert get_flash(conn, :error) == "The Test provider account is already bound to another user."
@@ -82,64 +91,42 @@ defmodule PowAssent.Phoenix.RegistrationControllerTest do
   end
 
   alias PowAssent.Test.EmailConfirmation.Phoenix.Endpoint, as: EmailConfirmationEndpoint
-  describe "GET /auth/:provider/create with PowEmailConfirmation" do
-    test "with user email", %{conn: conn} do
-      params = %{@provider => %{user_identity: %{"provider" => @provider, "uid" => "new_user", "token" => @token_params}, user: %{}}}
-      conn =
-        conn
-        |> Plug.Conn.put_session(:pow_assent_params, params)
-        |> Phoenix.ConnTest.dispatch(EmailConfirmationEndpoint, :post, Routes.pow_assent_registration_path(conn, :create, @provider), %{user: %{email: "foo@example.com"}})
+  alias PowAssent.Test.EmailConfirmation.Users.User, as: EmailConfirmationUser
+  describe "POST /auth/:provider/create with PowEmailConfirmation" do
+    @valid_params %{user: %{email: "foo@example.com"}}
+    @taken_params %{user: %{email: "taken@example.com"}}
+
+    test "with email from user", %{conn: conn} do
+      conn = Phoenix.ConnTest.dispatch(conn, EmailConfirmationEndpoint, :post, Routes.pow_assent_registration_path(conn, :create, @provider), @valid_params)
 
       refute Pow.Plug.current_user(conn)
 
       assert redirected_to(conn) == "/registration_created"
       assert get_flash(conn, :error) == "You'll need to confirm your e-mail before you can sign in. An e-mail confirmation link has been sent to you."
 
-      assert_received {:mail_mock, mail}
-      mail.html =~ "http://example.com/confirm-email/"
-    end
-
-    test "with provider email", %{conn: conn} do
-      params = %{@provider => %{user_identity: %{"provider" => @provider, "uid" => "new_user", "token" => @token_params}, user: %{"name" => "John Doe", "email" => "foo@example.com"}}}
-      conn =
-        conn
-        |> Plug.Conn.put_session(:pow_assent_params, params)
-        |> Phoenix.ConnTest.dispatch(EmailConfirmationEndpoint, :post, Routes.pow_assent_registration_path(conn, :create, @provider), %{user: %{}})
-
-      refute Pow.Plug.current_user(conn)
-
-      assert redirected_to(conn) == "/registration_created"
-      assert get_flash(conn, :error) == "You'll need to confirm your e-mail before you can sign in. An e-mail confirmation link has been sent to you."
+      assert user = Process.get({EmailConfirmationUser, :inserted})
+      assert user.email == "foo@example.com"
+      assert user.email_confirmation_token
+      refute user.email_confirmed_at
 
       assert_received {:mail_mock, mail}
       mail.html =~ "http://example.com/confirm-email/"
     end
 
-    test "with verified provider email", %{conn: conn} do
-      params = %{@provider => %{user_identity: %{"provider" => @provider, "uid" => "new_user", "token" => @token_params}, user: %{"name" => "John Doe", "email" => "foo@example.com", "email_verified" => true}}}
-      conn =
-        conn
-        |> Plug.Conn.put_session(:pow_assent_params, params)
-        |> Phoenix.ConnTest.dispatch(EmailConfirmationEndpoint, :post, Routes.pow_assent_registration_path(conn, :create, @provider), %{user: %{}})
+    test "with already taken email", %{conn: conn} do
+      conn = Phoenix.ConnTest.dispatch(conn, EmailConfirmationEndpoint, :post, Routes.pow_assent_registration_path(conn, :create, @provider), @taken_params)
 
-      assert redirected_to(conn) == "/registration_created"
-      assert Pow.Plug.current_user(conn)
-
-      refute_received {:mail_mock, _mail}
+      assert html = html_response(conn, 200)
+      assert html =~ "<label for=\"user_email\">Email</label>"
+      assert html =~ "<input id=\"user_email\" name=\"user[email]\" type=\"text\" value=\"taken@example.com\">"
+      assert html =~ "<span class=\"help-block\">has already been taken</span>"
     end
   end
 
+  alias PowAssent.Test.WithAccessToken.Phoenix.Endpoint, as: WithAccessTokenEndpoint
   describe "POST /auth/:provider/create recording strategy params" do
-    setup %{conn: conn} do
-      user_identity_params = %{"provider" => @provider, "uid" => "new_user_with_access_token", "token" => @token_params}
-      provider_params      = %{@provider => %{user_identity: user_identity_params, user: @user_params}}
-      conn                 = Plug.Conn.put_session(conn, :pow_assent_params, provider_params)
-
-      {:ok, conn: conn}
-    end
-
     test "records", %{conn: conn} do
-      conn = post conn, Routes.pow_assent_registration_path(conn, :create, @provider), %{user: %{email: "foo@example.com"}}
+      conn = Phoenix.ConnTest.dispatch(conn, WithAccessTokenEndpoint, :post, Routes.pow_assent_registration_path(conn, :create, @provider), %{user: %{email: "foo@example.com"}})
 
       assert redirected_to(conn) == "/registration_created"
       assert user = Pow.Plug.current_user(conn)