+++ test/support/with_access_token/user.ex
@@ -1,0 +1,12 @@
+defmodule PowAssent.Test.WithAccessToken.Users.User do
+  @moduledoc false
+  use Ecto.Schema
+  use Pow.Ecto.Schema
+  use PowAssent.Ecto.Schema
+
+  schema "users" do
+    pow_user_fields()
+
+    timestamps()
+  end
+end